Move WKB and WKT around, added implicits
